Josh Big Edit
I removed the original XR Rig and replaced it with the newer xr origin (action based) which should allow for porting to different headsets easier and gives us more options as far as pre-baked functionality. I also turned gravity off for the chickens and made them grabbable. Finally I added turning and teleportation on for the player via controller input.